% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2config.R
\name{flat2config}
\alias{flat2config}
\title{Convert flat list representation of WOFOST crop/program config file}
\usage{
flat2config(config_flat)
}
\arguments{
\item{config_flat}{the configuartion as flat R list}
}
\value{
a list with the configuration
}
\description{
Convert flat list representation of WOFOST crop/program config file
}
