% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatlist.R
\name{flatlist}
\alias{flatlist}
\title{Flatten config list}
\usage{
flatlist(mylist)
}
\arguments{
\item{mylist}{R List with WOFOST crop/program config file content}
}
\value{
the flat (unnested) list
}
\description{
Flatten config list
}
