% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_sa_Wofost.R
\name{optim_sa_Wofost}
\alias{optim_sa_Wofost}
\title{WOFOST parameter optimisation}
\usage{
optim_sa_Wofost(parfiles, wofostbin = .GlobalEnv$wofostbin,
  wofostdir = .GlobalEnv$wofostdir, qualityfun = .GlobalEnv$qualityfun,
  start, lower, upper, control = list(nlimit = 1, maxgood = 1),
  maximization = T)
}
\arguments{
\item{parfiles}{character vector: which parfiles to manipulate}

\item{wofostbin}{wofost executable (under unixoid systems use "wine wofost.exe")}

\item{wofostdir}{path in which the wofost executable resides (without the actual execuatable name)}

\item{qualityfun}{the function to be evaluated after wofost run}

\item{start}{named numeric vector with the start values}

\item{lower}{named numeric vector with the lower bounds (names must be the names of the parameters to manipulate)}

\item{upper}{named numeric vector with the upper bounds}

\item{control}{the control list passed to optim_sa}

\item{maximization}{maximization instead of minimization of the}

\item{parameters}{named vector: the parameters with their values to set}

\item{parfiles}{chracter vectore with the parameter files to be manipulated}
}
\value{
The WOFOST output file
}
\description{
Optimisation of WOFOST crop and management parameters (for calibration). This wraps \code{runWofost} for using with \code{optimization::optim_sa}. Since runWofost needs some extra info that optim_sa does not include in the function call, some Global variables are set for this.
}
