% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWeatherWofost.R
\name{readWeatherWofost}
\alias{readWeatherWofost}
\title{Read WOFOST weather file}
\usage{
readWeatherWofost(files)
}
\arguments{
\item{filename}{a single year WOFOST weather file}
}
\value{
a list with elements "data" (the actual weather data) and "site" (lon, lat, alt, Ångström coefficients, label)
}
\description{
Read WOFOST weather file
}
