% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWofostOutput.R
\name{readWofostOutput}
\alias{readWofostOutput}
\title{Read WOFOST output  file}
\usage{
readWofostOutput(filename)
}
\value{
a list with all the Wofost outputs (detailed and summaries)
}
\description{
Read WOFOST output  file
}
