% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWofost.R
\name{runWofost}
\alias{runWofost}
\title{Run WOFOST simulation}
\usage{
runWofost(wofostdir = .GlobalEnv$wofostdir,
  wofostbin = .GlobalEnv$wofostbin)
}
\arguments{
\item{wofostdir}{path in which the wofost executable resides (without the actual execuatable name)}

\item{wofostbin}{wofost executable (under unixoid systems use "wine wofost.exe")}
}
\description{
Run WOFOST simulation
}
