\name{wofost-class}

\docType{class}


\alias{DailyWeather}
\alias{Rcpp_DailyWeather}
\alias{WofostCrop}
\alias{WofostCropParameters}
\alias{WofostCropParametersNPK}

\alias{WofostSoil}
\alias{WofostSoilParameters}
\alias{WofostSoilParametersNPK}

\alias{WofostControl}
\alias{WofostModel}
\alias{WofostOutput}
\alias{Rcpp_WofostCrop}
\alias{Rcpp_WofostCropParameters}
\alias{Rcpp_WofostCropParametersNPK}
\alias{Rcpp_WofostSoil}
\alias{Rcpp_WofostSoilParameters}
\alias{Rcpp_WofostSoilParametersNPK}

\alias{Rcpp_WofostControl}
\alias{Rcpp_WofostOutput}
\alias{Rcpp_WofostModel}


\alias{DailyWeather-class}
\alias{Rcpp_DailyWeather-class}

\alias{WofostCrop-class}
\alias{WofostCropParameters-class}
\alias{WofostCropParametersNPK-class}

\alias{WofostSoil-class}
\alias{WofostSoilParameters-class}
\alias{WofostSoilParametersNPK-class}
\alias{WofostControl-class}
\alias{WofostModel-class}
\alias{WofostOutput-class}

\alias{Rcpp_WofostCrop-class}
\alias{Rcpp_WofostCropParameters-class}
\alias{Rcpp_WofostCropParametersNPK-class}

\alias{Rcpp_WofostSoil-class}
\alias{Rcpp_WofostSoilParameters-class}
\alias{Rcpp_WofostSoilParametersNPK-class}
\alias{Rcpp_WofostControl-class}
\alias{Rcpp_WofostOutput-class}
\alias{Rcpp_WofostModel-class}

\title{WOFOST model classes}

\description{
These are the classes!
}

\section{Objects from the Class}{
Objects can be created with this function: \code{\link{wofost}}.
}


\keyword{classes}

