% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapWofostOptim.R
\name{wrapWofostOptim}
\alias{wrapWofostOptim}
\title{wrap WOFOST run for optimisation functions}
\usage{
wrapWofostOptim(parameters, parnames = .GlobalEnv$parnames,
  parfiles = .GlobalEnv$parfiles, wofostdir = .GlobalEnv$wofostdir,
  wofostbin = .GlobalEnv$wofostbin, qualityfun = .GlobalEnv$qualityfun)
}
\arguments{
\item{parameters}{named vector: the parameters with their values to set}

\item{parfiles}{character vector: which parfiles to manipulate}
}
\value{
The WOFOST output file
}
\description{
wrap WOFOST run for optimisation functions
}
