% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeatherWofost.R
\name{writeWeatherWofost}
\alias{writeWeatherWofost}
\title{Write WOFOST weather file(s)}
\usage{
writeWeatherWofost(met, targetfolder, station_number = 1,
  prefix = met$site$shortlab)
}
\description{
Write WOFOST weather file(s)
}
