% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWofostParfile.R
\name{writeWofostParfile}
\alias{writeWofostParfile}
\title{write WOFOST configuration file}
\usage{
writeWofostParfile(config, outfile, spc = F)
}
\arguments{
\item{config}{the configration (crop definition in the CROPD folder or Management/Global simulation config in the RUNIO dir)}

\item{outfile}{the output file to write to}

\item{spc}{(boolean) use spaces beween varname and assignment}
}
\description{
write WOFOST configuration file
}
