\name{Rwofost-package}

\alias{Rwofost-package}

\alias{Rwofost}

\docType{package}

\title{
\packageTitle{Rwofost}
}

\description{
This package provides a R interface to the WOFOST crop growth simulation model.

This is the first release. Please consider this version unstable. It needs a bit more work to simplify its use.  

More detailed documentation is also forthcoming, but there is ample general documentation available on-line. The documentation for the FORTRAN version is most applicable. 

The model is programmed in C++. It was derived from the original FORTRAN implementation. It passes the tests cases developed for the PCSE/python version. This suggests that you safely can use the model for "standard" computation of potential and water-limited production. 
}

